权限 list页面显示

+++ godot/src/action/Delete.java
@@ -1,5 +1,11 @@
 package action;
 
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
+
+import org.apache.struts2.ServletActionContext;
+
 import com.opensymphony.xwork2.ActionSupport;
 
 import dao.SiteDao;
@@ -8,6 +14,13 @@ public class Delete extends ActionSupport {
 	
 	private String id = "";
 	public String execute(){
+		HttpServletRequest request = ServletActionContext.getRequest();
+		HttpServletResponse response = ServletActionContext.getResponse();
+		HttpSession session = request.getSession();
+		if(!session.getAttribute("user").equals("admin2015")){
+			return "home";
+		}
+		
 		if(!id.equals("")){
 			SiteDao siteDao = new SiteDao();
 			SiteDao.deleteSite(id);
